ToDo/Tasks/serializers.py
from rest_framework import serializers
from .models import Task

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ['id', 'title', 'description', 'priority', 'date_of_task', 'active_task', 'slug','user']
        read_only_fields=['user']
