requirements.txt
alembic==1.1.0.dev0
amqp==5.0.9
app==0.0.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.0
attrs==19.3.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
boto3==1.24.7
botocore==1.27.7
Brlapi==0.7.0
cached-property==1.5.1
celery==5.2.3
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
class-registry==2.1.2
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==3.2.11
django-braces==1.15.0
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-filter==21.1
django-money==2.1.1
django-phonenumber-field==6.1.0
django-storages==1.12.3
django-taggit==2.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
example==0.1.0
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.12.0
filters==1.3.2
future==0.18.2
h11==0.9.0
httplib2==0.14.0
idna==3.3
importlib-metadata==1.5.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
jsonfield==3.1.0
jsonschema==3.2.0
kafka==1.3.5
kazam==1.4.5
keyring==18.0.1
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.0
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
phonenumbers==8.12.48
phonenumberslite==8.12.48
pi==0.1.2
Pillow==9.1.0
platformdirs==3.5.1
pluggy==1.5.0
prompt-toolkit==3.0.26
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.3
py==1.8.1
py-moneyed==1.2
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==8.3.3
python-apt==2.0.1+ubuntu0.20.4.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-decouple==3.6
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
redis==4.1.2
regex==2022.3.2
relay==0.0.8
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.16.0
SQLAlchemy==1.3.12
sqlparse==0.4.2
ssh-import-id==5.10
stripe==2.76.0
systemd-python==234
tblib==1.7.0
termcolor==1.1.0
texttable==1.6.2
toml==0.10.0
tomli==2.2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.8
usb-creator==0.3.7
uvicorn==0.11.3
vine==5.0.0
virtualenv==20.23.0
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==1.2.3
websockets==8.1
WoeUSB-ng==0.2.10.post2
wrapt==1.13.3
wsproto==0.15.0
wxPython==4.0.7
xkit==0.0.0
zipp==1.0.0
